Выделен модуль холста симулякра
