+++ src/data/requests.ts
@@ -48,3 +48,35 @@ export async function fetchProject(uuid: string): Promise<Project> {
             })
     )
 }
+
+/*    public runCalculations(): void {
+        const socket = new WebSocket(`${WS_URL}/project/${this.project.uid}/changes`)
+        socket.onmessage = (event: MessageEvent) => {
+            const message = plainToClass(WebSocketMessage, JSON.parse(event.data))
+            if (message.type == WSMessageType.Request) {
+                const message = new WebSocketMessage()
+                message.payload = JSON.stringify({ length: this.eventLoop.length() })
+                socket.send(JSON.stringify(message))
+                return
+            }
+            if (message.payload) {
+                const state = plainToClass(SimulacrumState, JSON.parse(message.payload))
+                this.eventLoop.enqueue(state)
+            }
+        }
+    }
+
+    protected fetchProjectInitState(): void {
+        fetch(`${REST_URL}/projects/${this.project.uid}/objects`)
+            .then(response => response.json())
+            .then(stateJson => {
+                const state = plainToClass(SimulacrumState, stateJson)
+                for (const objInfo of state.objects) {
+                    const object = new CanvasObject(objInfo)
+                    this.scene.add(object.instance)
+                    this.objects[objInfo.id] = object
+                }
+                this.eventLoop.enqueue(state)
+                this.applayStepChanges()
+            })
+    }*/
